Added movement constraints to ControlPoint and, consequently, made adjacency list ordered.
Added/modified methods to resize grid, make grid arbitrary the grid an arbitrary size, make the MappedImage itself change size, brighten the displayed BufferedImage, change the color of the control points, rescale an image to fit in the MappedImage while maintaining the aspect ratio, and generate sets of triangles to use for warping.

Also included MorphTools and Triangle classes to perform warping.